Move DeviceStatus and its helpers to separate class
